+++ src/config/chain/neutron.ts
@@ -28,7 +28,7 @@ export const NEUTRON_CHAIN_CONFIGS: ChainConfigs = {
         enabled: true,
       },
       gov: {
-        enabled: true,
+        enabled: false,
       },
     },
     gas: {
@@ -70,7 +70,7 @@ export const NEUTRON_CHAIN_CONFIGS: ChainConfigs = {
         enabled: false,
       },
       gov: {
-        enabled: true,
+        enabled: false,
       },
     },
     gas: {