+++ src/config/chain/terra.ts
@@ -30,7 +30,7 @@ export const TERRA_CHAIN_CONFIGS: ChainConfigs = {
       },
       gov: {
         enabled: true,
-        disableOpenProposal: true,
+        hideOpenProposal: true,
       },
     },
     gas: {
@@ -73,7 +73,7 @@ export const TERRA_CHAIN_CONFIGS: ChainConfigs = {
       },
       gov: {
         enabled: true,
-        disableOpenProposal: true,
+        hideOpenProposal: true,
       },
     },
     gas: {